move hardcoded values to StaticConfiguration file

+++ FiveDevicesOrleans/DeviceMessage.cs
@@ -3,7 +3,7 @@
     public class DeviceMessage
     {
         public string DeviceId { get; set; }
-        public decimal Temperature { get; set; }
+        public double Temperature { get; set; }
         public long TimeStamp { get; set; }
     }
 }
